First, the token is now in a file "token.txt" so I can open this repository to the public. Next, I create a UsefulFunction file, (this name is temporary) to create an addRole function. Moreover, I implemented an idea of A. Meyer. The idea is to add role to people automatically with her student number. Actually, there is a command "/role update" to give roles to students. It remains for me to add role after a verification of a student instead of execute command who will try to add role to all members of the Discord.

+++ src/main/java/fr/nathan/ugebot/events/WelcomeListener.java
@@ -1,6 +1,5 @@
 package fr.nathan.ugebot.events;
 
-import fr.nathan.ugebot.fonctions.Verification;
 import net.dv8tion.jda.api.EmbedBuilder;
 import net.dv8tion.jda.api.Permission;
 import net.dv8tion.jda.api.entities.MessageType;
@@ -98,7 +97,7 @@ public class WelcomeListener extends ListenerAdapter {
     @Override
     public void onMessageReceived(@NotNull MessageReceivedEvent e) {
         if (e.getChannel().getName().equals("verifyme")) {
-            if (!e.getMessage().getType().equals(MessageType.SLASH_COMMAND)){
+            if (!e.getMessage().getType().equals(MessageType.SLASH_COMMAND) && !e.getMember().hasPermission(Permission.ADMINISTRATOR)) {
                 e.getMessage().delete().queue();
             }
         }
